Add files via upload
This code is for developing an OTP (One-Time Password) verification system using a GUI Interface. The program generate a 6-digit OTP and send it to the user's email address for verification. Upon receiving the OTP, the user needs to enter it into the system for validation. If the entered OTP matches the generated OTP,the user gets access; otherwise, access should be denied.